requirements.txt
fastapi
uvicorn[standard]
duckdb
pandas
pyarrow
python-dotenv
requests
fastapi-cache2
pydantic